Xenakis's Free Stochastic Music program (Formalized Music 2e 1992, pp. 145-52) adapted to modern Fortran compiler
Xenakis's Free Stochastic Music program (Formalized Music 2e 1992, pp. 145-52) adapted for use with a  modern Fortran compiler. The code in FREE_STOCHASTIC_MUSIC.F has been tested in Approximatrix Simply Fortran 3.7 (https://simplyfortran.com/).
Relatively few changes have been made to the code as found in Formalized Music. Function names have been updated, following Simply Fortran's error prompts, in order for the code to run as formatted Fortran. For example, at target 230, line XEN 202, "X1=RAND(0)" replaces "X1=RANF(-1)" from the original code. Input and output devices and their associated functions have been added so that the input data may be read from the file DATA.TXT.
The data in DATA.TXT has been transcribed from Formalized Music, pp. 151-52 with few alterations. In line 25, space 52 (the digit preceding "12000" at the end of line 25), the "0" in the original input data has been replaced by "1" so that the program will produce different output on each run. This is in accordance with the comment on lines XEN 151-52 of the original code: "IF ALEA IS NON-ZERO EACH RUN OF THIS PROGRAM WILL PRODUCE DIFFERENT OUTPUT DATA." If the user desires to produce the same data each time, the "1" in line 25, space 52 may be replaced with the original "0". Line 27 of the original data is blank. Here the blank has been filled in with a string of nine zeroes, which ensure that the values for variables KTEST 3, 1, and 2 are each set to 0.